refactor(lsp): unify command detection via which
Re-lands commit e1ed463a70 of #780 on current develop, per the Neonforge98
audit. The lsp crate carried three copies of a "is this tool on PATH?"
probe, each spawning `which` / `where` with slightly different env
handling:

- commands/discovery.rs: `var("PATH").unwrap_or_default()` forwarded
  unconditionally, so an unset or non-UTF-8 PATH ran the child with
  `PATH=""`.
- lint_tools.rs and workspace_scan/process.rs: no env handling, the child
  inherited the parent's PATH as-is.

All three are replaced by one `command_detection::command_exists` that
resolves in-process with `which::which` (already a dependency, already
used by `install_pipeline::find_binary`), so the PATH subtlety the
original extraction changed silently no longer exists:

- PATH unset / empty: before, discovery spawned `which` with `PATH=""`
  (false unless the tool sat in cwd) while lint_tools/workspace_scan let
  `which` fall back to its own default search path; now every caller
  returns false.
- Non-UTF-8 PATH entries: before, discovery dropped the whole PATH
  (`var` errors, `unwrap_or_default` gives ""); now they are walked
  because `which` reads PATH with `var_os`.
- Windows: PATHEXT is honoured (`tsc` finds `tsc.cmd`), no child process
  and no console window.
- Names containing a path separator resolve relative to cwd, like the
  `which` binary; no caller passes one.

Reachability on develop (cdf5a3a35 made LSP/lint agent-only, no lsp/lint
Tauri command is registered in handler_list.inc):

- discovery.rs (`lsp_check_installed`), package_manager.rs and
  install_pipeline.rs are reachable through agent-core's manage_lsp tool
  and `LspManager::ensure_binary`.
- lint_tools.rs (`check_lint_tools` <- `lint_check_installed`),
  workspace_scan/mod.rs (`lint_scan_get_tools`), orchestrator.rs
  (`lint_scan_orchestrated`) and process.rs (`eslint_available`) have no
  caller anywhere in src-tauri. Their private `command_exists` variants
  are deleted here; the dead public functions themselves are only
  re-pointed at the shared helper because removing them cascades into the
  whole workspace_scan module and the lint-tool catalog, which is a
  separate dead-code PR.

`lsp::commands::command_exists` (a glob re-export of the discovery copy)
is no longer public; no crate imported it.

Adds unit tests for the helper (`sh` / `cmd` resolves, a random name and
an empty name do not).
