+++ src-tauri/crates/lsp/src/lib.rs
@@ -6,6 +6,7 @@
 //! Also includes ESLint integration for style/formatting diagnostics.
 
 pub mod codec;
+mod command_detection;
 pub mod commands;
 pub mod config;
 pub mod eslint;
